+++ next-docs/components/sidebar/navigation.ts
@@ -21,6 +21,7 @@ const navigation = [
     children: [
       { name: 'Accordion', href: '/components/accordion' },
       { name: 'Avatar', href: '/components/avatar' },
+      { name: 'AuthCode', href: '/components/authCode' },
       { name: 'Banner', href: '/components/banner' },
       { name: 'Button', href: '/components/button' },
       { name: 'Carousel', href: '/components/carousel' },