+++ next-docs/public/examples/avatar/Fallbacks.tsx
@@ -1,11 +1,10 @@
-import { Avatar } from '@heathmont/moon-components';
+import { Avatar } from '@heathmont/moon-core';
 
 const Example = () => (
   <div className="flex justify-around items-center w-full">
-    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="xsmall" />
-    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="small" />
-    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="medium" />
-    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="large" />
+    <Avatar color="gohan.100" backgroundColor="piccolo.100"/>
+    <Avatar color="gohan.100" backgroundColor="piccolo.100"/>
+    <Avatar color="gohan.100" backgroundColor="piccolo.100"/>
   </div>
 );