+++ next-docs/public/examples/heading/Uppercase.tsx
@@ -1,0 +1,10 @@
+import { Heading } from '@heathmont/moon-core';
+
+const Example = () => (
+  <div className="flex justify-around w-full items-center">
+    <Heading>Default text-transform is none</Heading>
+    <Heading isUppercase>Uppercase Heading</Heading>
+  </div>
+);
+
+export default Example;