+++ next-docs/public/examples/label/SizesList.tsx
@@ -5,27 +5,30 @@ import { ControlsClose, GenericInfo } from '@heathmont/moon-icons';
 const Example = () => (
   <div className="flex flex-col justify-around items-center w-full">
     <div className="flex justify-around items-center w-full mb-4">
-      <Label size="small">Small</Label>
-      <Label size="small" iconLeft={<GenericInfo />}>
-        Small
+      <Label size="twoxsmall">Twoxsmall</Label>
+      <Label size="twoxsmall" iconLeft={<GenericInfo color='hit.100' />}>
+        Twoxsmall
       </Label>
-      <Label size="small" iconRight={<ControlsClose />}>
-        Small
+      <Label size="twoxsmall" iconRight={<ControlsClose color='hit.100' />}>
+        Twoxsmall
       </Label>
       <Label
-        size="small"
-        iconLeft={<GenericInfo />}
-        iconRight={<ControlsClose />}
+        size="twoxsmall"
+        iconLeft={<GenericInfo color='hit.100' />}
+        iconRight={<ControlsClose color='hit.100' />}
       >
-        Small
+        Twoxsmall
       </Label>
     </div>
     <div className="flex justify-around items-center w-full">
-      <Label>Medium</Label>
-      <Label iconLeft={<GenericInfo />}>Medium</Label>
-      <Label iconRight={<ControlsClose />}>Medium</Label>
-      <Label iconLeft={<GenericInfo />} iconRight={<ControlsClose />}>
-        Medium
+      <Label>Xsmall</Label>
+      <Label iconLeft={<GenericInfo color='hit.100' />}>Xsmall</Label>
+      <Label iconRight={<ControlsClose color='hit.100' />}>Xsmall</Label>
+      <Label
+        iconLeft={<GenericInfo color='hit.100' />}
+        iconRight={<ControlsClose color='hit.100'/>}
+      >
+        Xsmall
       </Label>
     </div>
   </div>