+++ next-docs/public/examples/text/Colors.tsx
@@ -2,7 +2,7 @@ import { Text } from '@heathmont/moon-core';
 
 const Example = () => (
   <div className="flex justify-around w-full items-center">
-    <Text>Defaul color is Bulma</Text>
+    <Text>Default color is Bulma</Text>
     <Text color="piccolo.100">Piccolo</Text>
     <Text color="trunks.100">Trunks</Text>
     <Text color="krillin.100">Krillin</Text>