+++ next-docs/public/examples/text/Underline.tsx
@@ -1,0 +1,10 @@
+import { Text } from '@heathmont/moon-core';
+
+const Example = () => (
+  <div className="flex justify-around w-full items-center">
+    <Text>Default text-decoration is none</Text>
+    <Text isUnderline>Underline Text</Text>
+  </div>
+);
+
+export default Example;