+++ packages/components/src/datepicker/private/helpers/getPlaceholder.ts
@@ -1,5 +1,5 @@
-import isValid from 'date-fns/isValid';
 import format from 'date-fns/format';
+import isValid from 'date-fns/isValid';
 
 const DateFormatWithTime = 'dd/MM/yyyy HH:mm';
 const DateFormat = 'dd/MM/yyyy';