+++ packages/components/src/loader/Loader.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import styled, { css, keyframes } from 'styled-components';
-import { rem, themed } from '@heathmont/moon-utils';
 import { ColorProps } from '@heathmont/moon-themes';
+import { rem, themed } from '@heathmont/moon-utils';
+import styled, { css, keyframes } from 'styled-components';
 
 /**
  * Animation