+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 8.4.0
+
+### Minor Changes
+
+- Clubhouse icon, lint, label, avatar, auth code, typography
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@8.4.0
+  - @heathmont/moon-utils@8.4.0
+
 ## 8.3.7
 
 ### Patch Changes